requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
alembic>=1.13.0
SQLAlchemy>=2.0.25
supabase>=2.0.0

# Redis & Caching
redis[hiredis]>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bleach>=6.1.0

# HTTP Client
httpx[http2]>=0.26.0

# Performance
orjson>=3.9.10
msgpack>=1.0.7

# Monitoring & Logging
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0

# AI/ML
openai>=1.10.0
anthropic>=0.18.0
# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0

# Code Execution & Security
RestrictedPython>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
